+++ lib/utility.php
@@ -507,7 +507,9 @@ function push_out_data_input_method($data_input_id) {
  * @param int $poller_id - the poller_id of the buffer
  */
 function poller_update_poller_cache_from_buffer($local_data_ids, &$poller_items, $poller_id = 1) {
-	if ($poller_id > 1) {
+	global $config;
+
+	if ($poller_id > 1 && $config['poller_id'] == 1) {
 		$rcnn_id = poller_connect_to_remote($poller_id);
 	}
 
@@ -521,7 +523,7 @@ function poller_update_poller_cache_from_buffer($local_data_ids, &$poller_items,
 				SET present=0
 				WHERE local_data_id IN ($ids)");
 
-			if ($poller_id > 1) {
+			if ($poller_id > 1 && $config['poller_id'] == 1 && $rcnn_id !== false) {
 				db_execute("UPDATE poller_item
 					SET present=0
 					WHERE local_data_id IN ($ids)", true, $rcnn_id);
@@ -577,7 +579,7 @@ function poller_update_poller_cache_from_buffer($local_data_ids, &$poller_items,
 			if ($overhead + $buf_len > $max_packet - 1024) {
 				db_execute($sql_prefix . $buffer . $sql_suffix);
 
-				if ($poller_id > 1) {
+				if ($poller_id > 1 && $config['poller_id'] == 1 && $rcnn_id !== false) {
 					db_execute($sql_prefix . $buffer . $sql_suffix, true, $rcnn_id);
 				}
 
@@ -593,7 +595,7 @@ function poller_update_poller_cache_from_buffer($local_data_ids, &$poller_items,
 	if ($buf_count > 0) {
 		db_execute($sql_prefix . $buffer . $sql_suffix);
 
-		if ($poller_id > 1) {
+		if ($poller_id > 1 && $config['poller_id'] == 1 && $rcnn_id !== false) {
 			db_execute($sql_prefix . $buffer . $sql_suffix, true, $rcnn_id);
 		}
 	}
@@ -604,7 +606,7 @@ function poller_update_poller_cache_from_buffer($local_data_ids, &$poller_items,
 			WHERE present=0
 			AND local_data_id IN ($ids)");
 
-		if ($poller_id > 1) {
+		if ($poller_id > 1 && $config['poller_id'] == 1 && $rcnn_id !== false) {
 			db_execute("DELETE FROM poller_item
 				WHERE present=0
 				AND local_data_id IN ($ids)", true, $rcnn_id);
